requirements.txt
python-dotenv
pyTelegramBotAPI
psycopg2